Allow construction even when the provided buffer is not large enough to hold the entire inner payload
For protocol layers where an inner payload length is provided in the header (IP and UDP), allow construction even when the provided buffer is not large enough to hold the entire inner payload, as long as the headers are in tact. This allows this library to be used to parse packet samples where payloads have been intentionally truncated.