Ignore further bytes in the TCP header after End-of-Option-List is encountered
I've come across the following options attached to TCP headers in the wild:

[00 00 08 00]
[00 00 10 00]

Before this commit, these were being rejected as invalid because of the non-zero byte following the End-of-Option-List option. It is ambiguous as to whether RFC793 specifies that "header padding" bytes should be zero. Nevertheless Linux ignores all header bytes after an End-of-Options-List is encountered, so we will too.